MOFI/Backend/User/LastB/requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
motor
pymongo
dnspython
python-dotenv
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
email-validator
itsdangerous
emails
httpx==0.24.1
fastapi-authlib
alembic
sqlalchemy
greenlet
cffi
cryptography
pyjwt
authlib
dynaconf
cloudinary